Changed Edit/Delete privileges to owners of microposts only.

+++ src/Controller/MicroPostController.php
@@ -7,13 +7,16 @@ use App\Entity\MicroPost;
 use App\Repository\MicroPostRepository;
 use App\Form\MicroPostType;
 use Doctrine\ORM\EntityManagerInterface;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Session\Flash\FlashBagInterface;
+use Symfony\Component\HttpKernel\Exception\UnauthorizedHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\Routing\RouterInterface;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 
 /**
  * @Route("micro-post")
@@ -35,6 +38,10 @@ class MicroPostController
      * @var FlashBagInterface
      */
     private $flashBag;
+    /**
+     * @var AuthorizationCheckerInterface
+     */
+    private $authorizationChecker;
 
     public function __construct(
         \Twig\Environment $twig,
@@ -42,7 +49,8 @@ class MicroPostController
         FormFactoryInterface $formFactory,
         EntityManagerInterface $entityManager,
         RouterInterface $router,
-        FlashBagInterface $flashBag
+        FlashBagInterface $flashBag,
+        AuthorizationCheckerInterface $authorizationChecker
     ) {
         $this->twig = $twig;
         $this->microPostRepository = $microPostRepository;
@@ -50,6 +58,7 @@ class MicroPostController
         $this->entityManager = $entityManager;
         $this->router = $router;
         $this->flashBag = $flashBag;
+        $this->authorizationChecker = $authorizationChecker;
     }
 
     /**
@@ -67,33 +76,44 @@ class MicroPostController
 
     /**
      * @Route("/edit/{id}", name="micro_post_edit")
+     * @Security("is_granted('edit', microPost)", message="Access Denied")
      */
-    public function edit(MicroPost $microPost, Request $request)
+    //(Way-3 (Above))
+    public function edit(MicroPost $microPost, Request $request/*, (Way-1)AuthorizationCheckerInterface $authorizationChecker*/)
     {
-        $form = $this->formFactory->create(MicroPostType::class, $microPost);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->entityManager->persist($microPost);
-            $this->entityManager->flush();
-
-            return new RedirectResponse($this->router->generate("micro_post_index"));
+        //$this->denyUnlessGranted("edit", $microPost); --> Requires Controller extension.
+        /* (Way-2)
+        if (!$this->authorizationChecker->isGranted("edit", $microPost)) {
+            throw new UnauthorizedHttpException();
         }
+        */
+            $form = $this->formFactory->create(MicroPostType::class, $microPost);
+            $form->handleRequest($request);
 
-        $html = $this->twig->render(
-            "micro-post/add.html.twig",
-            ["form" => $form->createView()]
-        );
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->entityManager->persist($microPost);
+                $this->entityManager->flush();
 
+                return new RedirectResponse($this->router->generate("micro_post_index"));
+            }
+
+            $html = $this->twig->render(
+                "micro-post/add.html.twig",
+                ["form" => $form->createView()]
+            );
         return new Response($html);
     }
 
 
     /**
      * @Route("/delete/{id}", name="micro_post_delete")
+     * @Security("is_granted('delete', microPost)", message="Access Denied")
      */
     public function delete(MicroPost $microPost)
     {
+        if (!$this->authorizationChecker->isGranted("edit", $microPost)) {
+            throw new UnauthorizedHttpException();
+        }
         $this->entityManager->remove($microPost);
         $this->entityManager->flush();
